pyproject.toml
[project]
name = "reminder"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "plyer>=2.1.0",
    "apscheduler>=3.10.0",
    "python-dateutil>=2.8.2",  # Для месячного интервала
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
]
